drivers: timer: mcux_gpt_timer: Add tracing code to MCUX GPT timer
This commit adds a rudimentary tracing system to the GPT timer. It is
useful when debugging subtle timing issues within the driver.

The current state of the tracing system is that the code should be
compiled with debug optimizations, and then the tracing buffer defined
in the timer driver can be dumped in order to backtrace test failures
created by the timer driver.

Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>