soc: rt10xx: Move low power mode exit to power_state_exit function
low power mode exit code should be in power_state_exit function to be
compatible with power subsystem. Move clock changes and voltage increase
to this function.

Note that the power management subsystem is NOT called from a direct GPT
isr, hence why this function was not used before. When the GPT isr is
converted to a normal indirect ISR, this function will be called on the
first GPT interrupt that wakes the system from sleep.

Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>