soc: rt10xx: Only set GPT interrupt as active in GPC when required
LPM sample in SDK only sets the GPT interrupt as enabled in the general
power controller module (which controls transition to and from WAIT
state) when the device is about to sleep, and disables the GPT interrupt
in the GPC after that. Update LPM code in zephyr to follow this pattern.

Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>