enabled loading itcm and SDRAM samples via MCUBoot
enabled loading itcm and sdram code samples using MCUBoot ram loader.

Keep in mind to do this, the base image must be signed with IMGTOOL, and
a header size of 0x400 should be used.

This is to keep in line with the base address adjustments of 0x400
applied by these overlays. The code itself will have a base address offset of
0x400 when loaded, since the header is also copied to SDRAM. Thus the
base address change is required. Note also that MCUBoot must have
platform specific init enabled for ITCM loading, since the image is over
128KB.

Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>