scripts: build: remove KEEP() from symbols moved with zephyr_code_relocate
The build script parsing files that will be relocated using the
zephyr_code_relocate macro was adding KEEP() to all symbols.

This resulted in unnecessary symbols being added to the build, and could
result in a linker failure if symbols in the relocated file that the
linker would otherwise discard referenced undefined symbols.

This error was observed while relocating portions of the networking
subsystem to SRAM, since the network management API was not being compiled
but some unused functions in the network subsystem were using the API.

Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>