Revert "lib: libc: newlib: make sure retargetable locking is enabled in toolchain"
This reverts commit d486bd3cfe0b4bbdcbc08be2fa325f876046243a.

The newlib's Configure option [1] used by the ARM GCC toolchain 10.3
does not allow locking routine to be retargeted [2]. This is because
--enable-newlib-retargetable-locking was not set when the toolchain was
built. Newer versions have fixed this issues but it is not the right
time for us to update the toolchain since we want to stay in sync with
what mcux-sdk uses.

This patch is temporary and shall be reverted once the toolchain issue
is fixed.

[1] https://community.arm.com/support-forums/f/compilers-and-libraries-forum/53310/gcc-arm-none-eabi-what-were-the-newlib-compilation-options
[2] https://sourceware.org/legacy-ml/newlib/2016/msg01165.html

Signed-off-by: Stephane Viau <stephane.viau@nxp.com>