boards: add mcuboot tag to all boards capable of building MCUBoot test
Add mcuboot supported tag to all boards that can build the MCUBoot test in
tree. This list was created via the following steps:

- used twister to run tests/boot/test_mcuboot sample against all in tree
  boards
- parsed twister.json output report to determine which boards passed a
  build test
- added mcuboot tag to these boards

No on device testing has been performed at this time to verify these
platforms work. The goal of this commit is to simply serve as a starting
point recording which platforms may support MCUBoot with Zephyr.

Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>