dfu-usb: build to load images to RAM
Build using ramload for MCUBoot. This will load USB DFU sample to OCRAM
region, so that any bugs stemming from FlexSPI read-while-write can be
avoided. Note that MCUBoot boots the image with the newest version in
RAMLOAD mode, so once the hello world image is flashed (version 0.0.1),
the image will need to be erased to revert to the USB DFU image.

Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>