drivers: serial: enable low power mode support for Flexcomm UART
Enable low power mode support for Flexcomm UART driver. This support
enables the driver to set power mode contraints while transmitting, so
characters will not be dropped.

The UART can also be configured to wake the SOC. See below for the required
devicetree properties to enable this support, when CONFIG_PM=y:

&flexcomm0 {
	clocks = <&clkctl1 MCUX_OSC_32K_CLK>;
	current-speed = <9600>;
};

Note that the baud rate may be any value that is an exact divisor of
9600, IE 4800 or 2400. The `clocks` DT property must be set to the 32KHz
oscillator to enable this functionality.

Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>