testbed: added basic tracing code to debug issue 58729
Adding basic tracing code to debug issue 58729, primarily to the system
tick driver and kernel timeout code. This code places events in a trace
buffer. The failing test case will run a "bkpt" instruction when the
failure condition occurs, so the user can catch the chip in a debugger
at the point of failure.

The debugger can then be used (along with dump_samples) to produce a
somewhat readable markdown table of the event log up to the point of
failure.

Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>